(refactor) Validation 필요한 도메인으로 이동
