+++ src/main/java/step02/user/Name.java
@@ -1,11 +1,19 @@
 package step02.user;
 
+import step02.exception.OutOfNameLengthException;
+
 import java.util.Objects;
 
 public class Name {
+    private final static String ERROR_NAME_LENGTH = "이름은 %d~%d글자 이내로 입력해주세요";
+    private final static int MAX_NAME_LENGTH = 5;
+    private final static int MIN_NAME_LENGTH = 1;
+    private final static int ZERO = 0;
+
     private final String name;
 
-    public Name(String name) {
+    public Name(String name) throws OutOfNameLengthException {
+        checkNameLength(name);
         this.name = name;
     }
 
@@ -15,6 +23,12 @@ public class Name {
 
     public int getNameLength() {
         return name.length();
+    }
+
+    public static void checkNameLength(String name) throws OutOfNameLengthException {
+        if (name.length() <= ZERO || name.length() > MAX_NAME_LENGTH) {
+            throw new OutOfNameLengthException(String.format(ERROR_NAME_LENGTH, MIN_NAME_LENGTH, MAX_NAME_LENGTH));
+        }
     }
 
     @Override
